---
🖼️✨ **Добавлена система рендеринга и визуальные компоненты**
🎯 Введён интерфейс `Renderable` для универсального подхода к отрисовке
🧠 Используется композиция с `PositionComponent` вместо наследования
🛠️ Реализована `RenderSystem`, которая вызывает `render()` у всех `Renderable`
🟥 Добавлен `RectangleComponent` — визуальный примитив для отладки и интерфейса
🔌 Система легко расширяется под спрайты, текст, коллизии и т. д.

---

+++ src/main/java/pyc/lopatuxin/core/components/movement/PositionComponent.java
@@ -16,6 +16,6 @@ import pyc.lopatuxin.core.ecs.Component;
 @AllArgsConstructor
 public class PositionComponent implements Component {
 
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 }
